Added an alert to confirm the plant was added
GUOWEN, when you're integrating this into the server, you will see in the POST request I used "test" as the ID : value. You can change that to userID, in order to see if it works with the server, I just used "test" as a placeholder for testing purposes. It is on line 134.